move _get_auth() to AWSAuth class
